feat(@eurora/prosemirror-tauri-bindings): add a processor for editor state to string
