+++ apps/desktop/package.json
@@ -32,6 +32,7 @@
 		"@eurora/ext-video": "workspace:*",
 		"@eurora/ext-transcript": "workspace:*",
 		"@icons-pack/svelte-simple-icons": "^5.0.0",
+		"@eurora/prosemirror-tauri-bindings": "workspace:*",
 		"@eurora/typescript-config": "workspace:*",
 		"uuid": "^11.1.0",
 		"@playwright/test": "^1.52.0",
