--- apps/desktop/src/lib/query-processor.ts
@@ -1,23 +1,0 @@
-import { Editor } from '@eurora/prosemirror-core';
-import { Fragment } from 'prosemirror-model';
-function handleFragment(content: Fragment) {
-    if    
-}
-
-export function processQuery(editorRef: Editor): string {
-	const view = editorRef.view;
-	if (!view) {
-		throw new Error('No view found');
-	}
-	const doc = view.state.doc;
-	let query = '';
-	doc.content.forEach((node) => {
-		if (node.type == 'text') {
-			query += node.text;
-		}
-		if (node.type.name === 'video') {
-			console.log(node.attrs);
-		}
-	});
-	return query;
-}
