+++ packages/prosemirror/tauri-bindings/package.json
@@ -6,21 +6,9 @@
 	"module": "./dist/index.js",
 	"exports": {
 		".": {
-			"types": "./src/index.ts",
-			"svelte": "./src/index.ts",
+			"types": "./src/index.js",
+			"svelte": "./src/index.js",
 			"default": "./dist/index.js"
-		},
-		"./register-extensions": {
-			"types": "./src/register-extensions.ts",
-			"default": "./dist/register-extensions.js"
-		},
-		"./utils": {
-			"types": "./src/utils.ts",
-			"default": "./dist/utils.js"
-		},
-		"./example-usage": {
-			"types": "./src/example-usage.ts",
-			"default": "./dist/example-usage.js"
 		}
 	},
 	"files": [
